# swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Medilogue Swagger RestFul API UI",
      "description": "Node.js RestFul API UI",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080"
      }
    ],
    "paths": {
...
      "/calendar/:date": {
        "get": {
          "tags": ["calendar"],
          "description": "유저의 이번달 일정을 제공하는 API",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "schema": {
                "type": "string (JWT token)"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "example": {
                    "success": true,
                    "schedules": [
                        {
                            "content": "당뇨 관리 상담",
                            "startTime": "2023-02-15T00:00:00Z",
                            "attendee": "user2"
                        },
                        {
                            "content": "고혈압 관리 상담",
                            "startTime": "2023-02-15T01:00:00Z",
                            "attendee": "user1"
                        },
                        {
                            "content": "당뇨 관리 상담",
                            "startTime": "2023-02-18T01:00:00Z",
                            "attendee": "user2"
                        }
                    ]
                  }
                }
              }
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "calendar/update/{id}": {
        "put": {
          "tags": ["calendar"],
          "description": "mission의 object id req.param.id로 보내서 mission을 completed 상태로 업데이트하는 API",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "403": {
              "description": "Forbidden access, 권한이 없는 유저가 접근했을 때"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "mission/update/{id}": {
        "put": {
          "tags":["mission"],
          "description": "mission의 object id req.param.id로 보내서 mission을 completed 상태로 업데이트하는 API",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Mission not found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "in": "header",
          "bearerFormat": "JWT"
        }
      }
    }
  }